Refactor i_reblock and i_topology_utils
